requirements.txt
flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
openai==0.28.1
werkzeug==2.3.7
gunicorn==21.2.0
